refactor: split tools.js in smaller files
