+++ lib/tools.js
@@ -1,853 +1,26 @@
-const cds = require('@sap/cds')
-const z = require('zod')
+// Barrel module: re-exports the tool APIs from the split tool modules.
+// The implementation lives in ./tools/{describe,query,action}.js
+// with shared helpers in ./utils/tools-shared.js
 
+const { getInstructions } = require('./utils/tools-shared')
 const {
-  createReadInputSchema,
-  extractFieldsFromWhere,
-  createActionInputSchema,
-  createPerActionInputSchema,
-  getDescription,
-  resolveI18n,
-  parseAssertRange,
-  resolveTypeDef
-} = require('./utils/cds-to-schema')
-const { resolveQueryLimits } = require('./utils/limits')
-const { validateCqnTargets } = require('./utils/validate-targets')
-
-const LOG = cds.log('mcp')
-const { inspect } = require('util')
-const fmt = (obj) => inspect(obj, { depth: 8, compact: 3, breakLength: 80, colors: true })
-
-const formatResult = cds.env.mcp?.toon_format === false ? JSON.stringify : formatAsToon
-
-function formatAsToon(data) {
-  try {
-    var toon = (formatAsToon.toon ??= require('@toon-format/toon'))
-  } catch {
-    throw new Error(
-      '@toon-format/toon is not installed.\n' +
-        'Please install it with: npm add @toon-format/toon\n' +
-        'Or disable TOON format with: cds.env.mcp.toon_format = false'
-    )
-  }
-  return toon.encode(data)
-}
-
-// Draft-related elements added by CAP for draft-enabled entities
-const DRAFT_ELEMENTS = [
-  'IsActiveEntity',
-  'HasDraftEntity',
-  'HasActiveEntity',
-  'DraftAdministrativeData',
-  'DraftAdministrativeData_DraftUUID',
-  'SiblingEntity',
-  'DraftMessages'
-]
-
-const LOCALIZED_ELEMENTS = ['localized', 'texts']
-
-function getInstructions(def, locale, prefix_) {
-  locale = locale || cds.context?.locale || 'en'
-  const custom = resolveI18n(def['@mcp.instructions'], locale)
-  return (
-    custom ||
-    `Always use the \`${prefix_}describe\` tool to explore the data model and available actions/functions. ` +
-      `Only then use \`${prefix_}query\` to read data or \`${prefix_}call\` to invoke actions or functions.`
-  )
-}
-
-// Create MCP error response
-function errorResponse(message) {
-  return {
-    content: [{ type: 'text', text: message }],
-    isError: true
-  }
-}
-
-// Format error for display — includes details when CAP groups multiple errors
-function formatError(err) {
-  if (err.details?.length) {
-    return JSON.stringify(
-      err.details.map((d) => {
-        const e = { message: d.message }
-        if (d.target) e.target = d.target
-        return e
-      })
-    )
-  }
-  return err.message
-}
-
-// Build clean args object for logging (filter out empty values)
-function buildQueryArgs(args) {
-  return Object.fromEntries(
-    Object.entries(args).filter(
-      ([, v]) => v !== undefined && v !== null && !(Array.isArray(v) && v.length === 0)
-    )
-  )
-}
-
-// Validates field references including path expressions
-function validateFields(fields, entity, definitions) {
-  if (!fields || !Array.isArray(fields) || fields.length === 0) return []
-
-  return fields.filter((field) => {
-    // Skip aggregate/func expressions and other non-ref entries
-    if (!field || typeof field !== 'object' || !Array.isArray(field.ref)) return false
-
-    const segments = field.ref
-    let currentEntity = entity
-
-    for (let i = 0; i < segments.length; i++) {
-      const segment = segments[i]
-      // Extract field name from plain string or { id } object (infix filter segment)
-      const fieldName = typeof segment === 'string' ? segment : segment?.id
-      if (!fieldName) return true // Invalid - no field name
-
-      const element = currentEntity.elements?.[fieldName]
-
-      if (!element) return true // Invalid - element doesn't exist
-      if (element['@cds.api.ignore']) return true // Invalid - element is @cds.api.ignore
-
-      // If not the last segment, must be an association
-      if (i < segments.length - 1) {
-        if (!element.target) return true // Invalid - not an association
-
-        // Resolve association target
-        currentEntity = definitions[element.target]
-        if (!currentEntity) return true // Invalid - target entity not found
-      }
-    }
-
-    return false // Valid
-  })
-}
-
-function createGenericReadToolDefinition(entityNames, serviceName, prefix = '') {
-  const name = prefix + 'query'
-  if (cds.env.mcp?.format === 'cqn') {
-    return {
-      name,
-      description: `Query any entity in ${serviceName} service. Use describe to discover available entities and their fields.`,
-      inputSchema: createReadInputSchema({ entityNames }),
-      annotations: {
-        readOnlyHint: true,
-        idempotentHint: true,
-        openWorldHint: false
-      }
-    }
-  }
-
-  return {
-    name,
-    description:
-      `Query any entity in ${serviceName} service.` +
-      "Ensure to first use the `describe` tool to discover an entity's available fields.",
-    inputSchema: z.object({
-      cql: z
-        .string()
-        .describe(
-          'CAP CQL statement to execute queries; only SELECT statements are allowed. ' +
-            'CQL is a superset of SQL, that supports ' +
-            '- path expressions to follow associations instead of JOINs, e.g. SELECT author.name from Books, ' +
-            '- postfix projection syntax, with nested expands, e.g. SELECT from Authors { ID, name, books { ID, title }}, ' +
-            '- standard CAP/OData functions and aggregates (count, sum, avg, min, max, lower, upper, substring, year, month, day, etc.). ' +
-            'Database functions (CURRENT_USER, SESSION_USER, SYSUUID, CURRENT_SCHEMA) are not supported. ' +
-            'LIMIT is auto-injected to the service max.'
-        )
-    }),
-    annotations: {
-      readOnlyHint: true,
-      idempotentHint: true,
-      openWorldHint: false
-    }
-  }
-}
-
-function createDescribeToolDefinition(entityNames, actionNames, serviceName, prefix = '') {
-  const schemaFields = {}
-
-  if (entityNames.length > 0) {
-    schemaFields.entities = z
-      .array(z.enum(entityNames))
-      .optional()
-      .describe('Specific entities to get element details for.')
-  }
-
-  if (actionNames.length > 0) {
-    schemaFields.actions = z
-      .array(z.enum(actionNames))
-      .optional()
-      .describe('Specific actions or functions to get parameter details for.')
-  }
-
-  const name = prefix + 'describe'
-  return {
-    name,
-    description:
-      `Describe the data model of ${serviceName} service. ` +
-      `Returns an overview of all entities and actions with descriptions. ` +
-      `Specify 'entity' to get element details, or 'action' to get parameter details.`,
-    inputSchema: z.object(schemaFields),
-    annotations: {
-      readOnlyHint: true,
-      destructiveHint: false,
-      idempotentHint: true
-    }
-  }
-}
-
-function createCallActionToolDefinition(actionNames, serviceName, prefix = '') {
-  const name = prefix + 'call'
-  return {
-    name,
-    description: `Call an unbound action or function in ${serviceName} service. Use describe to discover available actions and their parameters.`,
-    inputSchema: createActionInputSchema(actionNames),
-    annotations: {
-      readOnlyHint: false,
-      destructiveHint: true,
-      idempotentHint: false,
-      openWorldHint: false
-    }
-  }
-}
-
-// Register a single generic query tool for all entities (default behavior)
-function registerGenericReadTool(server, srv, entities, prefix, { log = LOG } = {}) {
-  const entityNames = Object.keys(entities)
-  if (entityNames.length === 0) {
-    log.debug('No entities to register tools for', { service: srv.name })
-    return
-  }
-
-  const def = createGenericReadToolDefinition(entityNames, srv.name, prefix)
-
-  server.registerTool(
-    def.name,
-    {
-      description: def.description,
-      inputSchema: def.inputSchema,
-      annotations: def.annotations
-    },
-    (args) => executeGenericReadTool(srv, entities, args, { log })
-  )
-
-  log.debug('Registered generic tool', { tool: def.name, service: srv.name })
-}
-
-// Register the describe tool for service introspection
-function registerDescribeTool(server, srv, entities, actions = {}, prefix, { log = LOG } = {}) {
-  const entityNames = Object.keys(entities)
-  const actionNames = Object.keys(actions)
-  if (entityNames.length === 0 && actionNames.length === 0) {
-    log.debug('No entities or actions to describe', { service: srv.name })
-    return
-  }
-
-  const def = createDescribeToolDefinition(entityNames, actionNames, srv.name, prefix)
-
-  server.registerTool(
-    def.name,
-    {
-      description: def.description,
-      inputSchema: def.inputSchema,
-      annotations: def.annotations
-    },
-    (args) => executeDescribe(srv, entities, actions, args, { log })
-  )
-
-  log.debug('Registered tool', { tool: def.name, service: srv.name })
-}
-
-// Register the call tool for invoking unbound actions/functions
-function registerCallActionTool(server, srv, actions, prefix, { log = LOG } = {}) {
-  const actionNames = Object.keys(actions)
-  if (actionNames.length === 0) return // No actions to register
-
-  const def = createCallActionToolDefinition(actionNames, srv.name, prefix)
-
-  server.registerTool(
-    def.name,
-    {
-      description: def.description,
-      inputSchema: def.inputSchema,
-      annotations: def.annotations
-    },
-    (args) => executeCallActionTool(srv, actions, args, { log })
-  )
-
-  log.debug('Registered tool', { tool: def.name, service: srv.name, actions: actionNames })
-}
-
-const removeCdsPrefix = (type) => type?.replace(/^cds\./, '') || type
-
-// Resolve a type reference to a human-readable string, recursively flattening custom types
-function renderTypeRef(ref, model, depth = 0) {
-  if (!ref || depth > 5) return 'unknown'
-
-  // Array case (many / array of)
-  if (ref.items) {
-    return `Array of ${renderTypeRef(ref.items, model, depth + 1)}`
-  }
-
-  // Inline struct case
-  if (ref.elements) {
-    const fields = Object.entries(ref.elements)
-      .map(([name, el]) => `${name}: ${renderTypeRef(el, model, depth + 1)}`)
-      .join(', ')
-    return `{${fields}}`
-  }
-
-  // Type reference
-  if (ref.type) {
-    // CDS primitive
-    if (ref.type.startsWith('cds.')) return removeCdsPrefix(ref.type)
-
-    // Custom type — resolve from model
-    const typeDef = model && resolveTypeDef(ref.type, model)
-    if (typeDef) {
-      // Structured custom type
-      if (typeDef.elements) {
-        const fields = Object.entries(typeDef.elements)
-          .map(([name, el]) => `${name}: ${renderTypeRef(el, model, depth + 1)}`)
-          .join(', ')
-        return `{${fields}}`
-      }
-      // Scalar alias — recurse into the aliased type
-      if (typeDef.type) {
-        return renderTypeRef(typeDef, model, depth + 1)
-      }
-    }
-
-    // Fallback — unknown custom type, keep its name
-    return removeCdsPrefix(ref.type)
-  }
-
-  return 'unknown'
-}
-
-// Build a human-readable return type string from an action definition
-function describeReturns(action, model) {
-  if (!action.returns) return null
-  return renderTypeRef(action.returns, model)
-}
-
-// Per-action tool definition factory (shared with compile.js)
-function createPerActionToolDefinition(actionName, action, serviceName, model, prefix = '') {
-  let description = getDescription(action) || `Call ${action.kind} ${actionName} in ${serviceName}`
-
-  // Append return type info so LLMs know without calling describe
-  const returnsInfo = describeReturns(action, model)
-  if (returnsInfo) {
-    description += `. Returns: ${returnsInfo}`
-  }
-
-  const name = prefix + actionName
-  return {
-    name,
-    description,
-    inputSchema: createPerActionInputSchema(action, model),
-    annotations: {
-      readOnlyHint: action.kind === 'function',
-      destructiveHint: action.kind === 'action',
-      idempotentHint: action.kind === 'function',
-      openWorldHint: false
-    }
-  }
-}
-
-// Register individual tools per action/function
-
-function registerPerActionTools(server, srv, actions, prefix, { log = LOG } = {}) {
-  const actionEntries = Object.entries(actions)
-  if (actionEntries.length === 0) {
-    log.debug('No actions to register tools for', { service: srv.name })
-    return
-  }
-
-  for (const [actionName, action] of actionEntries) {
-    const def = createPerActionToolDefinition(actionName, action, srv.name, srv.model, prefix)
-
-    server.registerTool(
-      def.name,
-      {
-        description: def.description,
-        inputSchema: def.inputSchema,
-        annotations: def.annotations
-      },
-      (args) => executePerActionTool(srv, actionName, action, args, { log })
-    )
-
-    log.debug('Registered tool', { tool: def.name, service: srv.name, kind: action.kind })
-  }
-}
-
-// Execute a per-action tool (args ARE the parameters directly)
-async function executePerActionTool(srv, actionName, action, args, { log = LOG } = {}) {
-  log(actionName, fmt({ service: srv.name, ...buildQueryArgs(args) }))
-
-  try {
-    const result = await srv.send(actionName, args)
-
-    const structured = {
-      action: actionName,
-      kind: action.kind,
-      result
-    }
-
-    return {
-      content: [{ type: 'text', text: formatResult(structured) }],
-      structuredContent: structured
-    }
-  } catch (err) {
-    log.error(`Executing action '${actionName}' failed with: \n\n`, err)
-
-    if (err.code === 401 || err.code === 403) {
-      return errorResponse(
-        `Authorization error (${err.code}): Not authorized to call ${actionName}. ${formatError(err)}`
-      )
-    }
-
-    return errorResponse(`Error calling ${actionName}: ${formatError(err)}`)
-  }
-}
-
-async function executeGenericReadTool(srv, entities, args, { log = LOG } = {}) {
-  if (cds.env.mcp?.format === 'cqn') {
-    return _executeGenericReadCqn(srv, entities, args, { log })
-  }
-  return _executeGenericReadCql(srv, entities, args, { log })
-}
-
-// Default CQL length cap. Override via cds.env.mcp.cql.maxLength.
-const DEFAULT_CQL_MAX_LENGTH = 10_000
-
-// Run a pre-built CQN: validate + LIMIT + srv.run.
-// Returns { ok:true, data, count, one } or { ok:false, code, reason, entity, err }.
-async function _executeCqnQuery({ srv, entities, cqn, entityDef }) {
-  const definitions = srv.model?.definitions || {}
-  const allowedEntities = new Set(Object.keys(entities))
-
-  const validation = validateCqnTargets(cqn, {
-    allowedEntities,
-    serviceName: srv.name,
-    definitions,
-    rootEntity: entityDef
-  })
-  if (!validation.valid) {
-    return { ok: false, code: 'validation', reason: validation.reason, entity: validation.entity }
-  }
-
-  if (!cqn.SELECT.one && !cqn.SELECT.distinct) {
-    _enforceSelectLimit(cqn, entityDef, srv)
-  }
-
-  try {
-    const result = await srv.run(cqn)
-    const one = !!cqn.SELECT.one
-    if (one) return { ok: true, data: result || null, count: 1, one }
-    const resultArray = Array.isArray(result) ? result : [result]
-    return { ok: true, data: resultArray, count: result?.$count ?? resultArray.length, one }
-  } catch (err) {
-    return { ok: false, code: 'runtime', err }
-  }
-}
-
-// Map _executeCqnQuery failure to MCP error response.
-// ctx: { serviceName, targetLabel?, execPrefix }
-function _queryErrorResponse(res, ctx, log) {
-  if (res.code === 'validation') {
-    if (res.reason === 'function-not-allowed') {
-      return errorResponse(
-        `Error: ${res.entity} is not allowed. Only CAP-standard functions and aggregates are permitted.`
-      )
-    }
-    if (res.reason === 'pseudo-column-not-allowed') {
-      return errorResponse(
-        `Error: Reference '${res.entity}' is not allowed (database pseudo-column).`
-      )
-    }
-    if (res.reason === 'expand-not-allowed') {
-      return errorResponse(
-        `Error: Cannot expand into inaccessible entity via '${res.entity}'. The target entity is restricted.`
-      )
-    }
-    return errorResponse(
-      `Error: Entity '${res.entity}' cannot be resolved for service ${ctx.serviceName}. Use describe to see available entities.`
-    )
-  }
-
-  log?.error('Query execution failed', { target: ctx.targetLabel, error: formatError(res.err) })
-  if (res.err.code === 401 || res.err.code === 403) {
-    const scope = ctx.targetLabel ? `You are not authorized to read ${ctx.targetLabel}. ` : ''
-    return errorResponse(`Authorization error (${res.err.code}): ${scope}${formatError(res.err)}`)
-  }
-  return errorResponse(`${ctx.execPrefix}: ${formatError(res.err)}`)
-}
-
-// Inject/clamp LIMIT (entity → service → env)
-function _enforceSelectLimit(cqn, targetEntity, srv) {
-  const { default: defaultLimit, max } = resolveQueryLimits(targetEntity || {}, srv.definition)
-  const effectiveMax = max ?? Number.MAX_SAFE_INTEGER
-  const effectiveDefault = defaultLimit ?? effectiveMax
-
-  if (cqn.SELECT.limit === null) {
-    cqn.SELECT.limit = { rows: { val: effectiveDefault } }
-    return
-  }
-
-  const { rows } = cqn.SELECT.limit || (cqn.SELECT.limit = {})
-  const existingVal = rows?.val ?? rows
-  cqn.SELECT.limit.rows = {
-    val: existingVal ? Math.min(existingVal, effectiveMax) : effectiveDefault
-  }
-}
-
-// CQL mode: parse → shared core → CQL response
-async function _executeGenericReadCql(srv, entities, args, { log = LOG } = {}) {
-  const rawCql = args.cql ?? args.sql ?? ''
-
-  const maxLength = cds.env.mcp?.cql?.maxLength ?? DEFAULT_CQL_MAX_LENGTH
-  if (rawCql.length > maxLength) {
-    return errorResponse(
-      `Error: CQL exceeds max length of ${maxLength} characters (received ${rawCql.length}).`
-    )
-  }
-
-  // CQL parser treats newlines as terminators
-  const cql = rawCql.replace(/\n/g, ' ')
-  log('query', fmt({ service: srv.name, cql }))
-
-  let cqn
-  try {
-    cqn = cds.parse.cql(cql)
-  } catch (err) {
-    log.error('CQL parse failed', { cql, error: err.message })
-    return errorResponse(`Error executing CQL: ${err.message}`)
-  }
-
-  if (!cqn.SELECT) return errorResponse('Error: Only SELECT statements are allowed.')
-
-  // Resolve target for LIMIT + expand context
-  const targetName = cqn.SELECT.from?.ref?.[0]
-  const localName =
-    targetName && srv.name && targetName.startsWith(srv.name + '.')
-      ? targetName.slice(srv.name.length + 1)
-      : targetName
-  const entityDef = localName ? srv.entities?.[localName] : null
-
-  // Report $count independent of LIMIT
-  cqn.SELECT.count = true
-
-  const res = await _executeCqnQuery({ srv, entities, cqn, entityDef })
-
-  if (!res.ok) {
-    return _queryErrorResponse(
-      res,
-      { serviceName: srv.name, targetLabel: null, execPrefix: 'Error executing CQL' },
-      log
-    )
-  }
-
-  const structured = { count: res.count, data: res.data }
-  return {
-    content: [{ type: 'text', text: formatResult(structured) }],
-    structuredContent: structured
-  }
-}
-
-// CQN mode: build CQN → shared core → CQN response
-async function _executeGenericReadCqn(srv, entities, args, { log = LOG } = {}) {
-  const { entity: entityName } = args
-  log('query', fmt({ service: srv.name, ...buildQueryArgs(args) }))
-
-  const entity = entities[entityName]
-  const definitions = srv.model?.definitions || {}
-
-  // Fast-fail field checks for clearer errors
-  if (args.where && Array.isArray(args.where)) {
-    const referencedFields = extractFieldsFromWhere(args.where)
-    const invalid = validateFields(referencedFields, entity, definitions)
-    if (invalid.length > 0) {
-      const fieldNames = [...new Set(invalid.map((f) => f.ref.join('.')))]
-      return errorResponse(
-        `Error: Invalid where field(s): ${fieldNames.join(', ')}. Use describe to see available fields for ${entityName}.`
-      )
-    }
-  }
-
-  const invalidSelect = validateFields(args.select, entity, definitions)
-  if (invalidSelect.length > 0) {
-    const fieldNames = invalidSelect.map((f) => f.ref.join('.'))
-    return errorResponse(
-      `Error: Invalid select field(s): ${fieldNames.join(', ')}. Use describe to see available fields for ${entityName}.`
-    )
-  }
-
-  // Build CQN
-  let query = SELECT.from(entityName)
-
-  if (args.distinct) query.SELECT.distinct = true
-
-  if (args.select && Array.isArray(args.select) && args.select.length > 0) {
-    query.columns(...args.select)
-  }
-
-  if (args.where && Array.isArray(args.where) && args.where.length > 0) {
-    query.where(args.where)
-  }
-
-  if (args.search) query.search(args.search)
-
-  if (args.groupBy && Array.isArray(args.groupBy) && args.groupBy.length > 0) {
-    query.groupBy(...args.groupBy)
-  }
-
-  if (args.having && Array.isArray(args.having) && args.having.length > 0) {
-    query.having(args.having)
-  }
-
-  if (args.orderBy && Array.isArray(args.orderBy) && args.orderBy.length > 0) {
-    query.orderBy(...args.orderBy)
-  }
-
-  if (args.one) {
-    query.SELECT.one = true
-  } else if (args.distinct) {
-    query.SELECT.limit = null
-  } else if (args.limit || args.offset) {
-    // Raw values; _enforceSelectLimit clamps/defaults
-    query.limit(args.limit ?? 0, args.offset ?? 0)
-  }
-
-  const res = await _executeCqnQuery({ srv, entities, cqn: query, entityDef: entity })
-  if (!res.ok) {
-    return _queryErrorResponse(
-      res,
-      { serviceName: srv.name, targetLabel: entityName, execPrefix: `Error reading ${entityName}` },
-      log
-    )
-  }
-  const structured = { entity: entityName, count: res.count, data: res.data }
-  return {
-    content: [{ type: 'text', text: formatResult(structured) }],
-    structuredContent: structured
-  }
-}
-
-async function executeDescribe(srv, entities, actions, args, { log = LOG } = {}) {
-  return _executeDescribeCsn(srv, entities, actions, args, { log })
-}
-
-// CSN format (default): return JSON with element metadata
-async function _executeDescribeCsn(srv, entities, actions, args, { log = LOG } = {}) {
-  // Determine what to include:
-  // - If only entity specified => return only those entities (no actions)
-  // - If only action specified => return only those actions (no entities)
-  // - If both specified => return both
-  // - If neither specified => return all entities and all actions
-  const hasEntity = args.entities?.length > 0
-  const hasAction = args.actions?.length > 0
-  const includeEntities = hasEntity || !hasAction
-  const includeActions = hasAction || !hasEntity
-
-  // Detail mode: show elements/parameters when specific entity/action is requested
-  const isEntityDetail = hasEntity
-  const isActionDetail = hasAction
-
-  const entityNames = hasEntity ? args.entities : Object.keys(entities)
-  const actionNamesToDescribe = hasAction ? args.actions : Object.keys(actions || {})
-
-  log(
-    'describe',
-    fmt({
-      service: srv.name,
-      ...(includeEntities && { entities: entityNames }),
-      ...(includeActions && { actions: actionNamesToDescribe })
-    })
-  )
-
-  const description = {
-    service: srv.name,
-    description: getDescription(srv.definition) || `Service ${srv.name}`
-  }
-
-  // Add entities if included
-  if (includeEntities) {
-    description.entities = {}
-
-    for (const entityName of entityNames) {
-      const entity = entities[entityName]
-      if (!entity) continue
-
-      const entityDescription = getDescription(entity) || `Entity ${entityName}`
-
-      // Overview mode: only description
-      if (!isEntityDetail) {
-        description.entities[entityName] = { description: entityDescription }
-        continue
-      }
-
-      // Detail mode: include queryLimits and elements
-      const queryLimits = resolveQueryLimits(entity, srv.definition)
-
-      const keys = []
-      const elements = {}
-
-      for (const [elemName, elem] of Object.entries(entity.elements || {})) {
-        // Skip draft-related, localized, and @cds.api.ignore elements
-        if (DRAFT_ELEMENTS.includes(elemName) || LOCALIZED_ELEMENTS.includes(elemName)) continue
-        if (elem['@cds.api.ignore']) continue
-
-        if (elem.key) keys.push(elemName)
-
-        const elemDescription = getDescription(elem) || `Element ${elemName}`
-
-        if (elem.target) {
-          // Association element
-          elements[elemName] = {
-            type: `${removeCdsPrefix(elem.type)} (${elem.is2one ? '1-1' : '1-*'})`,
-            target: elem.target,
-            description: elemDescription
-          }
-        } else {
-          // Regular element
-          const elemOutput = {
-            type: removeCdsPrefix(elem.type),
-            description: elemDescription
-          }
-          if (elem.notNull || elem['@mandatory']) {
-            elemOutput.notNull = true
-          }
-          if (elem.enum) {
-            elemOutput.enum = Object.fromEntries(
-              Object.entries(elem.enum).map(([key, { val }]) => [key, val])
-            )
-          }
-          if (elem['@assert.range']) {
-            const range = parseAssertRange(elem['@assert.range'])
-            elemOutput.range = range?.text || elem['@assert.range']
-          }
-          if (elem['@assert.format']) {
-            elemOutput.format = elem['@assert.format']
-          }
-          elements[elemName] = elemOutput
-        }
-      }
-
-      description.entities[entityName] = {
-        description: entityDescription,
-        keys,
-        queryLimits,
-        elements
-      }
-    }
-  }
-
-  // Add actions if included
-  if (includeActions) {
-    description.actions = {}
-
-    for (const actionName of actionNamesToDescribe) {
-      const action = actions?.[actionName]
-      if (!action) continue
-      const actionDescription = getDescription(action) || `${action.kind} ${actionName}`
-
-      // Overview mode: only kind and description
-      if (!isActionDetail) {
-        description.actions[actionName] = {
-          kind: action.kind,
-          description: actionDescription
-        }
-        continue
-      }
-
-      // Detail mode: include parameters and returns
-      const returns = describeReturns(action, srv.model)
-
-      description.actions[actionName] = {
-        kind: action.kind,
-        description: actionDescription,
-        parameters: {},
-        returns
-      }
-
-      // Add parameter descriptions
-      for (const [paramName, param] of Object.entries(action.params || {})) {
-        const paramDescription = getDescription(param) || null
-        const paramOutput = {
-          type: renderTypeRef(param, srv.model),
-          notNull: param.notNull || param['@mandatory'] || false,
-          description: paramDescription
-        }
-        if (param.enum) {
-          paramOutput.enum = Object.fromEntries(
-            Object.entries(param.enum).map(([key, { val }]) => [key, val])
-          )
-        }
-        if (param['@assert.range']) {
-          const range = parseAssertRange(param['@assert.range'])
-          paramOutput.range = range?.text || param['@assert.range']
-        }
-        if (param['@assert.format']) {
-          paramOutput.format = param['@assert.format']
-        }
-        description.actions[actionName].parameters[paramName] = paramOutput
-      }
-    }
-  }
-
-  return {
-    content: [{ type: 'text', text: formatResult(description) }],
-    structuredContent: description
-  }
-}
-
-async function executeCallActionTool(srv, actions, args, { log = LOG } = {}) {
-  const { action: actionName, parameters = {} } = args
-  log(
-    'call',
-    fmt({ service: srv.name, action: actionName, parameters: buildQueryArgs(parameters) })
-  )
-
-  const action = actions[actionName]
-  if (!action) {
-    return errorResponse(
-      `Error: Action '${actionName}' not found in service. Use describe to see available actions.`
-    )
-  }
-
-  try {
-    // Call the action via srv.send()
-    const result = await srv.send(actionName, parameters)
-
-    // Deep-clone result to plain objects — CAP may return typed instances
-    // that TOON cannot enumerate (e.g. results from actions returning `many`)
-    const plainResult = result != null ? JSON.parse(JSON.stringify(result)) : result
-
-    const structured = {
-      action: actionName,
-      kind: action.kind, // 'action' or 'function'
-      result: plainResult
-    }
-
-    return {
-      content: [{ type: 'text', text: formatResult(structured) }],
-      structuredContent: structured
-    }
-  } catch (err) {
-    log.error(`Executing action '${actionName}' failed with: \n\n`, err)
-
-    // Handle authorization errors
-    if (err.code === 401 || err.code === 403) {
-      return errorResponse(
-        `Authorization error (${err.code}): Not authorized to call ${actionName}. ${formatError(err)}`
-      )
-    }
-
-    return errorResponse(`Error calling ${actionName}: ${formatError(err)}`)
-  }
-}
+  createDescribeToolDefinition,
+  registerDescribeTool,
+  executeDescribe
+} = require('./tools/describe')
+const {
+  createGenericReadToolDefinition,
+  registerGenericReadTool,
+  executeGenericReadTool
+} = require('./tools/query')
+const {
+  createCallActionToolDefinition,
+  createPerActionToolDefinition,
+  registerCallActionTool,
+  registerPerActionTools,
+  executeCallActionTool,
+  executePerActionTool
+} = require('./tools/action')
 
 module.exports = {
   // Definition factories (shared with compile.js)
